refactor: DB 엔티티 기본 키 타입 Int -> Long 변경 (#12)

+++ core/database/src/main/java/com/example/database/entity/AssetEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "assets")
 data class AssetEntity(
     @PrimaryKey(autoGenerate = true)
-    val assetId: Int = 0,
+    val assetId: Long = 0,
 
     val assetType: AssetType,         // MODEL / DECORATION / BACKGROUND / TEXT_TEMPLATE
     val unityKey: String,             // Addressable key or prefab name
